refactor: add `Layout` component to the folder of the same name
